+++ log-base/CHANGELOG.md
@@ -1,3 +1,6 @@
+# log-base-0.8.0.0 (2019-04-02)
+* Add `getLoggerEnv` function to `MonadLog` class, add `getLoggerIO` utility.
+
 # log-base-0.7.4.0 (2017-10-27)
 * Add `mkBulkLogger'` ([#40](https://github.com/scrive/log/pull/40).